requirements.txt
PyYAML
tqdm
numpy
scipy
pandas
pillow
matplotlib
scikit-learn
torch
torchvision
fastapi[all]
pydantic
python-multipart
dvc[gdrive]
